# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BioPay – Multi-Tenant Subsurface Palm Vein Biometric Payment System
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface MessageResponse {
  message: string;
}

export interface UserRegistration {
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface MerchantRegistration {
  shop_name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginCredentials {
  email: string;
  password: string;
}

export type AuthResponseRole =
  (typeof AuthResponseRole)[keyof typeof AuthResponseRole];

export const AuthResponseRole = {
  user: "user",
  merchant: "merchant",
  admin: "admin",
} as const;

export interface AuthResponse {
  token: string;
  role: AuthResponseRole;
  id: number;
  /** @nullable */
  name?: string | null;
  email?: string;
}

export type SessionInfoRole =
  (typeof SessionInfoRole)[keyof typeof SessionInfoRole];

export const SessionInfoRole = {
  user: "user",
  merchant: "merchant",
  admin: "admin",
} as const;

export interface SessionInfo {
  id: number;
  role: SessionInfoRole;
  email: string;
  /** @nullable */
  name?: string | null;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  wallet_balance: number;
  is_verified: boolean;
  biometric_enrolled: boolean;
  created_at?: string;
}

export interface WalletBalance {
  wallet_balance: number;
  /** @nullable */
  message?: string | null;
}

export interface BiometricEnrollment {
  /** SHA-256 hash of palm vein scan data */
  biometric_hash: string;
}

export interface MerchantProfile {
  id: number;
  shop_name: string;
  email: string;
  merchant_balance: number;
  kiosk_id: string;
  is_online?: boolean;
  created_at?: string;
}

export interface EarningsSummary {
  today_earnings: number;
  today_count: number;
  total_earnings: number;
  total_count: number;
  weekly_earnings: number;
}

export interface PaymentInitiation {
  /** @minimum 1 */
  amount: number;
}

export type PaymentSessionStatus =
  (typeof PaymentSessionStatus)[keyof typeof PaymentSessionStatus];

export const PaymentSessionStatus = {
  WAITING: "WAITING",
  COMPLETED: "COMPLETED",
  FAILED: "FAILED",
  CANCELLED: "CANCELLED",
} as const;

export interface PaymentSession {
  session_id: string;
  amount: number;
  status: PaymentSessionStatus;
  /** @nullable */
  message?: string | null;
}

export type TransactionType =
  (typeof TransactionType)[keyof typeof TransactionType];

export const TransactionType = {
  TOPUP: "TOPUP",
  PURCHASE: "PURCHASE",
} as const;

export type TransactionStatus =
  (typeof TransactionStatus)[keyof typeof TransactionStatus];

export const TransactionStatus = {
  SUCCESS: "SUCCESS",
  FAILED: "FAILED",
} as const;

export interface Transaction {
  id: number;
  /** @nullable */
  user_id?: number | null;
  /** @nullable */
  merchant_id?: number | null;
  amount: number;
  type: TransactionType;
  status: TransactionStatus;
  /** @nullable */
  razorpay_order_id?: string | null;
  /** @nullable */
  user_name?: string | null;
  /** @nullable */
  merchant_name?: string | null;
  timestamp: string;
}

export interface TransactionPage {
  items: Transaction[];
  total: number;
  page: number;
  limit: number;
}

export interface TopupOrderInput {
  /**
   * Amount in INR
   * @minimum 10
   */
  amount: number;
}

export interface TopupOrder {
  razorpay_order_id: string;
  /** Amount in paise */
  amount: number;
  currency: string;
  /** Razorpay key_id for frontend checkout */
  key: string;
}

export interface TopupVerification {
  razorpay_order_id: string;
  razorpay_payment_id: string;
  razorpay_signature: string;
}

export interface AdminStats {
  total_users: number;
  total_merchants: number;
  total_volume: number;
  total_transactions: number;
  active_kiosks: number;
  failed_transactions: number;
  today_volume?: number;
  today_transactions?: number;
}

export interface AdminUser {
  id: number;
  name: string;
  email: string;
  wallet_balance: number;
  is_verified: boolean;
  biometric_enrolled: boolean;
  created_at: string;
}

export interface AdminUserPage {
  items: AdminUser[];
  total: number;
  page: number;
  limit: number;
}

export interface AdminMerchant {
  id: number;
  shop_name: string;
  email: string;
  merchant_balance: number;
  kiosk_id: string;
  is_online: boolean;
  created_at: string;
}

export interface AdminMerchantPage {
  items: AdminMerchant[];
  total: number;
  page: number;
  limit: number;
}

export interface KioskStatus {
  kiosk_id: string;
  merchant_id: number;
  shop_name: string;
  is_online: boolean;
  /** @nullable */
  last_seen?: string | null;
}

export interface BiometricScanInput {
  biometric_hash: string;
  merchant_id: number;
  amount: number;
}

export interface PaymentResult {
  success: boolean;
  transaction_id: number;
  amount: number;
  user_name: string;
  merchant_name: string;
  new_balance: number;
}

export interface HeartbeatInput {
  kiosk_id: string;
}

export type GetUserTransactionsParams = {
  page?: number;
  limit?: number;
};

export type GetMerchantTransactionsParams = {
  page?: number;
  limit?: number;
};

export type GetAdminUsersParams = {
  page?: number;
  limit?: number;
  search?: string;
};

export type GetAdminMerchantsParams = {
  page?: number;
  limit?: number;
  search?: string;
};

export type GetAdminTransactionsParams = {
  page?: number;
  limit?: number;
  type?: GetAdminTransactionsType;
};

export type GetAdminTransactionsType =
  (typeof GetAdminTransactionsType)[keyof typeof GetAdminTransactionsType];

export const GetAdminTransactionsType = {
  TOPUP: "TOPUP",
  PURCHASE: "PURCHASE",
} as const;