# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BioPay – Multi-Tenant Subsurface Palm Vein Biometric Payment System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register/user:
    post:
      operationId: registerUser
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRegistration"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/register/merchant:
    post:
      operationId: registerMerchant
      tags: [auth]
      summary: Register a new merchant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MerchantRegistration"
      responses:
        "201":
          description: Merchant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login/user:
    post:
      operationId: loginUser
      tags: [auth]
      summary: Login as user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login/merchant:
    post:
      operationId: loginMerchant
      tags: [auth]
      summary: Login as merchant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login/admin:
    post:
      operationId: loginAdmin
      tags: [auth]
      summary: Login as admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginCredentials"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated session info
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionInfo"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── User Dashboard ────────────────────────────────────────────────────────
...
  /merchants/me/transactions:
    get:
      operationId: getMerchantTransactions
      tags: [merchants]
      summary: Get merchant transaction history
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPage"
...
  /merchants/me/pos/initiate:
    post:
      operationId: initiatePayment
      tags: [merchants]
      summary: Initiate a payment — signals Raspberry Pi to scan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInitiation"
      responses:
        "200":
          description: Payment initiated, Pi signaled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSession"
...
  /hardware/verify-scan:
    post:
      operationId: verifyBiometricScan
      tags: [hardware]
      summary: Raspberry Pi submits biometric scan result to process payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BiometricScanInput"
      responses:
        "200":
          description: Payment processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResult"
        "402":
          description: Insufficient funds
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /hardware/heartbeat:
    post:
      operationId: hardwareHeartbeat
      tags: [hardware]
      summary: Kiosk heartbeat to report online status
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HeartbeatInput"
      responses:
        "200":
          description: Acknowledged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── Auth Schemas ────────────────────────────────────────────────────────
    UserRegistration:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    MerchantRegistration:
      type: object
      required: [shop_name, email, password]
      properties:
        shop_name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginCredentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, role, id]
      properties:
        token:
          type: string
        role:
          type: string
          enum: [user, merchant, admin]
        id:
          type: integer
        name:
          type: ["string", "null"]
        email:
          type: string
    SessionInfo:
      type: object
      required: [id, role, email]
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, merchant, admin]
        email:
          type: string
        name:
          type: ["string", "null"]

    # ── User Schemas ────────────────────────────────────────────────────────
...
    PaymentInitiation:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          minimum: 1
    PaymentSession:
      type: object
      required: [session_id, amount, status]
      properties:
        session_id:
          type: string
        amount:
          type: number
        status:
          type: string
          enum: [WAITING, COMPLETED, FAILED, CANCELLED]
        message:
          type: ["string", "null"]

    # ── Transaction Schemas ─────────────────────────────────────────────────
    Transaction:
      type: object
      required: [id, amount, type, status, timestamp]
      properties:
        id:
          type: integer
        user_id:
          type: ["integer", "null"]
        merchant_id:
          type: ["integer", "null"]
        amount:
          type: number
        type:
          type: string
          enum: [TOPUP, PURCHASE]
        status:
          type: string
          enum: [SUCCESS, FAILED]
        razorpay_order_id:
          type: ["string", "null"]
        user_name:
          type: ["string", "null"]
        merchant_name:
          type: ["string", "null"]
        timestamp:
          type: string
          format: date-time
    TransactionPage:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # ── Wallet Top-up Schemas ───────────────────────────────────────────────
...
    BiometricScanInput:
      type: object
      required: [biometric_hash, merchant_id, amount]
      properties:
        biometric_hash:
          type: string
        merchant_id:
          type: integer
        amount:
          type: number
    PaymentResult:
      type: object
      required: [success, transaction_id, amount, user_name, merchant_name, new_balance]
      properties:
        success:
          type: boolean
        transaction_id:
          type: integer
        amount:
          type: number
        user_name:
          type: string
        merchant_name:
          type: string
        new_balance:
          type: number
    HeartbeatInput:
      type: object
      required: [kiosk_id]
      properties:
        kiosk_id:
          type: string