# lib/api-zod/src/generated/types/loginCredentials.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BioPay – Multi-Tenant Subsurface Palm Vein Biometric Payment System
 * OpenAPI spec version: 0.1.0
 */

export interface LoginCredentials {
  email: string;
  password: string;
}