# lib/api-zod/src/generated/types/paymentSession.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BioPay – Multi-Tenant Subsurface Palm Vein Biometric Payment System
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentSessionStatus } from "./paymentSessionStatus";

export interface PaymentSession {
  session_id: string;
  amount: number;
  status: PaymentSessionStatus;
  /** @nullable */
  message?: string | null;
}