# lib/api-zod/src/generated/types/sessionInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BioPay – Multi-Tenant Subsurface Palm Vein Biometric Payment System
 * OpenAPI spec version: 0.1.0
 */
import type { SessionInfoRole } from "./sessionInfoRole";

export interface SessionInfo {
  id: number;
  role: SessionInfoRole;
  email: string;
  /** @nullable */
  name?: string | null;
}